+++ backend/lib/core/index.ts
@@ -1,2 +1,2 @@
-export * from './helper.js';
 export * from './server.js';
+export * from './forwarder.js';
